backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.7.0
python-dotenv==1.2.1
python-multipart==0.0.20

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0

# Redis
redis==5.2.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pyjwt==2.8.0
cryptography==42.0.5
svix==1.24.0
requests==2.31.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0

